+++ packages/ui-react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog @tonconnect/ui-react
 
+## 2.4.4
+
+### Patch Changes
+
+- Updated dependencies [fb9f6ca]
+    - @tonconnect/ui@2.4.4
+
+## 2.4.3
+
+### Patch Changes
+
+- Updated dependencies [39fa11d]
+    - @tonconnect/ui@2.4.3
+
+## 2.4.2
+
+### Patch Changes
+
+- Updated dependencies [b9d0441]
+    - @tonconnect/ui@2.4.2
+
 ## 2.4.1
 
 ### Patch Changes