+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog @tonconnect/ui
 
+## 2.4.4
+
+### Patch Changes
+
+- fb9f6ca: Fixed `tapps/launchParams` sessionStorage fallback not working because the value is a raw
+  URL query string, not JSON
+
+## 2.4.3
+
+### Patch Changes
+
+- 39fa11d: fix(ui): add launch params fallback
+
+## 2.4.2
+
+### Patch Changes
+
+- b9d0441: Remove unsupported in old browsers iterator helpers
+
 ## 2.4.1
 
 ### Patch Changes